#82 int() добавлен в def _calculate_total_amount(), в class LoanCreateview();

+++ pawnshop_crm/pawnshop/views/loan_views.py
@@ -63,7 +63,7 @@ class LoanCreateView(CreateView):
         return super().form_valid(form)
 
     def _calculate_total_amount(self, client_amount, duration, interest_rate):
-        total_amount = 100 * client_amount / (99 - interest_rate * duration)
+        total_amount = int(100 * client_amount / (99 - interest_rate * duration))
         return total_amount
 
     def _get_date_of_expire(self, duration):
